fixed pgen compatibility issue
