--- src/engines/engine_init_impl.hpp
@@ -1,67 +1,0 @@
-#ifndef ENGINES_ENGINE_INIT_IMPL_HPP
-#define ENGINES_ENGINE_INIT_IMPL_HPP
-
-#include "enums.h"
-#include "global.h"
-
-#include "archetypes/field_setter.h"
-#include "archetypes/traits.h"
-
-#include <Kokkos_Core.hpp>
-
-#include <string>
-
-namespace ntt {
-
-  template <SimEngine::type S, class M>
-    requires IsCompatibleWithEngine<S, M>
-  void Engine<S, M>::init() {
-    m_metadomain.InitStatsWriter(m_params, is_resuming);
-#if defined(OUTPUT_ENABLED)
-    m_metadomain.InitWriter(&m_adios, m_params);
-    m_metadomain.InitCheckpointWriter(&m_adios, m_params);
-#endif
-    logger::Checkpoint("Initializing Engine", HERE);
-    if (not is_resuming) {
-      // start a new simulation with initial conditions
-      logger::Checkpoint("Loading initial conditions", HERE);
-      if constexpr (arch::traits::pgen::HasInitFlds<user::PGen<S, M>>) {
-        logger::Checkpoint("Initializing fields from problem generator", HERE);
-        m_metadomain.runOnLocalDomains([&](auto& loc_dom) {
-          Kokkos::parallel_for(
-            "InitFields",
-            loc_dom.mesh.rangeActiveCells(),
-            arch::SetEMFields_kernel<decltype(m_pgen.init_flds), S, M> {
-              loc_dom.fields.em,
-              m_pgen.init_flds,
-              loc_dom.mesh.metric });
-        });
-      }
-      if constexpr (
-        arch::traits::pgen::HasInitPrtls<user::PGen<S, M>, Domain<S, M>>) {
-        logger::Checkpoint("Initializing particles from problem generator", HERE);
-        m_metadomain.runOnLocalDomains([&](auto& loc_dom) {
-          m_pgen.InitPrtls(loc_dom);
-        });
-      }
-    } else {
-#if defined(OUTPUT_ENABLED)
-      // read simulation data from the checkpoint
-      raise::ErrorIf(
-        m_params.template get<timestep_t>("checkpoint.start_step") == 0,
-        "Resuming simulation from a checkpoint requires a valid start_step",
-        HERE);
-      logger::Checkpoint("Resuming simulation from a checkpoint", HERE);
-      m_metadomain.ContinueFromCheckpoint(&m_adios, m_params);
-#else
-      raise::Error(
-        "Resuming simulation from a checkpoint requires -D output=ON",
-        HERE);
-#endif
-    }
-    print_report();
-  }
-
-} // namespace ntt
-
-#endif // ENGINES_ENGINE_INIT_IMPL_HPP
