+++ src/kernels/fields_bcs.hpp
@@ -44,18 +44,20 @@ namespace kernel::bc {
              metric::traits::HasConvert<M> &&
              ((S == SimEngine::SRPIC && metric::traits::HasTransform_i<M>) ||
               S == SimEngine::GRPIC) &&
-             (S == SimEngine::SRPIC && (traits::fieldsetter::HasEx1<I, M::Dim> ||
-                                        traits::fieldsetter::HasEx2<I, M::Dim> ||
-                                        traits::fieldsetter::HasEx3<I, M::Dim> ||
-                                        traits::fieldsetter::HasBx1<I, M::Dim> ||
-                                        traits::fieldsetter::HasBx2<I, M::Dim> ||
-                                        traits::fieldsetter::HasBx3<I, M::Dim>) ||
-              (S == SimEngine::GRPIC && (traits::fieldsetter::HasDx1<I, M::Dim> ||
-                                         traits::fieldsetter::HasDx2<I, M::Dim> ||
-                                         traits::fieldsetter::HasDx3<I, M::Dim>) ||
-               traits::fieldsetter::HasBx1<I, M::Dim> ||
-               traits::fieldsetter::HasBx2<I, M::Dim> ||
-               traits::fieldsetter::HasBx3<I, M::Dim>))
+             (S == SimEngine::SRPIC &&
+                (::traits::fieldsetter::HasEx1<I, M::Dim> ||
+                 ::traits::fieldsetter::HasEx2<I, M::Dim> ||
+                 ::traits::fieldsetter::HasEx3<I, M::Dim> ||
+                 ::traits::fieldsetter::HasBx1<I, M::Dim> ||
+                 ::traits::fieldsetter::HasBx2<I, M::Dim> ||
+                 ::traits::fieldsetter::HasBx3<I, M::Dim>) ||
+              (S == SimEngine::GRPIC &&
+                 (::traits::fieldsetter::HasDx1<I, M::Dim> ||
+                  ::traits::fieldsetter::HasDx2<I, M::Dim> ||
+                  ::traits::fieldsetter::HasDx3<I, M::Dim>) ||
+               ::traits::fieldsetter::HasBx1<I, M::Dim> ||
+               ::traits::fieldsetter::HasBx2<I, M::Dim> ||
+               ::traits::fieldsetter::HasBx3<I, M::Dim>))
   struct MatchBoundaries_kernel {
     static_assert(static_cast<dim_t>(o) < static_cast<dim_t>(M::Dim),
                   "Invalid component index");
@@ -123,12 +125,12 @@ namespace kernel::bc {
           metric.template convert<Crd::Cd, Crd::Ph>({ i1_ }, x_Ph_0);
           metric.template convert<Crd::Cd, Crd::Ph>({ i1_ + HALF }, x_Ph_H);
 
-          if constexpr (traits::fieldsetter::HasEx1<I, Dim::_1D> or
-                        traits::fieldsetter::HasBx2<I, Dim::_1D> or
-                        traits::fieldsetter::HasBx3<I, Dim::_1D>) {
+          if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_1D> or
+                        ::traits::fieldsetter::HasBx2<I, Dim::_1D> or
+                        ::traits::fieldsetter::HasBx3<I, Dim::_1D>) {
             const auto s = shape(math::abs(
               metric.template convert<i, Crd::Cd, Crd::Ph>(i1_ + HALF) - xg_edge));
-            if constexpr (traits::fieldsetter::HasEx1<I, Dim::_1D>) {
+            if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_1D>) {
               if (tags & BC::E) {
                 Fld(i1, em::ex1) = s * Fld(i1, em::ex1) +
                                    (ONE - s) *
@@ -138,17 +140,17 @@ namespace kernel::bc {
               }
             }
             // if constexpr (defines_bx2 or defines_bx3) {
-            if constexpr (traits::fieldsetter::HasBx2<I, Dim::_1D> or
-                          traits::fieldsetter::HasBx3<I, Dim::_1D>) {
+            if constexpr (::traits::fieldsetter::HasBx2<I, Dim::_1D> or
+                          ::traits::fieldsetter::HasBx3<I, Dim::_1D>) {
               if (tags & BC::B) {
-                if constexpr (traits::fieldsetter::HasBx2<I, Dim::_1D>) {
+                if constexpr (::traits::fieldsetter::HasBx2<I, Dim::_1D>) {
                   Fld(i1, em::bx2) = s * Fld(i1, em::bx2) +
                                      (ONE - s) *
                                        metric.template transform<2, Idx::T, Idx::U>(
                                          { i1_ + HALF },
                                          fset.bx2(x_Ph_H));
                 }
-                if constexpr (traits::fieldsetter::HasBx3<I, Dim::_1D>) {
+                if constexpr (::traits::fieldsetter::HasBx3<I, Dim::_1D>) {
                   Fld(i1, em::bx3) = s * Fld(i1, em::bx3) +
                                      (ONE - s) *
                                        metric.template transform<3, Idx::T, Idx::U>(
@@ -158,12 +160,12 @@ namespace kernel::bc {
               }
             }
           }
-          if constexpr (traits::fieldsetter::HasBx1<I, Dim::_1D> or
-                        traits::fieldsetter::HasEx2<I, Dim::_1D> or
-                        traits::fieldsetter::HasEx3<I, Dim::_1D>) {
+          if constexpr (::traits::fieldsetter::HasBx1<I, Dim::_1D> or
+                        ::traits::fieldsetter::HasEx2<I, Dim::_1D> or
+                        ::traits::fieldsetter::HasEx3<I, Dim::_1D>) {
             const auto s = shape(math::abs(
               metric.template convert<i, Crd::Cd, Crd::Ph>(i1_) - xg_edge));
-            if constexpr (traits::fieldsetter::HasBx1<I, Dim::_1D>) {
+            if constexpr (::traits::fieldsetter::HasBx1<I, Dim::_1D>) {
               if (tags & BC::B) {
                 Fld(i1, em::bx1) = s * Fld(i1, em::bx1) +
                                    (ONE - s) *
@@ -173,17 +175,17 @@ namespace kernel::bc {
               }
             }
             // if constexpr (defines_ex2 or defines_ex3) {
-            if constexpr (traits::fieldsetter::HasEx2<I, Dim::_1D> or
-                          traits::fieldsetter::HasEx3<I, Dim::_1D>) {
+            if constexpr (::traits::fieldsetter::HasEx2<I, Dim::_1D> or
+                          ::traits::fieldsetter::HasEx3<I, Dim::_1D>) {
               if (tags & BC::E) {
-                if constexpr (traits::fieldsetter::HasEx2<I, Dim::_1D>) {
+                if constexpr (::traits::fieldsetter::HasEx2<I, Dim::_1D>) {
                   Fld(i1, em::ex2) = s * Fld(i1, em::ex2) +
                                      (ONE - s) *
                                        metric.template transform<2, Idx::T, Idx::U>(
                                          { i1_ },
                                          fset.ex2(x_Ph_0));
                 }
-                if constexpr (traits::fieldsetter::HasEx3<I, Dim::_1D>) {
+                if constexpr (::traits::fieldsetter::HasEx3<I, Dim::_1D>) {
                   Fld(i1, em::ex3) = s * Fld(i1, em::ex3) +
                                      (ONE - s) *
                                        metric.template transform<3, Idx::T, Idx::U>(
@@ -210,9 +212,9 @@ namespace kernel::bc {
         const auto i2_ = COORD(i2);
 
         // SRPIC
-        if constexpr (traits::fieldsetter::HasEx1<I, Dim::_2D> or
-                      traits::fieldsetter::HasDx1<I, Dim::_2D> or
-                      traits::fieldsetter::HasBx2<I, Dim::_2D>) {
+        if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_2D> or
+                      ::traits::fieldsetter::HasDx1<I, Dim::_2D> or
+                      ::traits::fieldsetter::HasBx2<I, Dim::_2D>) {
           // i1 + 1/2, i2
           real_t xi_Cd;
           if constexpr (o == in::x1) {
@@ -227,17 +229,17 @@ namespace kernel::bc {
           coord_t<Dim::_2D> x_Ph_H0 { ZERO };
           metric.template convert<Crd::Cd, Crd::Ph>({ i1_ + HALF, i2_ }, x_Ph_H0);
 
-          if constexpr (traits::fieldsetter::HasEx1<I, Dim::_2D> or
-                        traits::fieldsetter::HasDx1<I, Dim::_2D>) {
+          if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_2D> or
+                        ::traits::fieldsetter::HasDx1<I, Dim::_2D>) {
             if ((tags & BC::E) or (tags & BC::D)) {
-              if constexpr (traits::fieldsetter::HasEx1<I, Dim::_2D> and
+              if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_2D> and
                             S == SimEngine::SRPIC) {
                 Fld(i1, i2, em::ex1) = s * Fld(i1, i2, em::ex1) +
                                        (ONE - s) *
                                          metric.template transform<1, Idx::T, Idx::U>(
                                            { i1_ + HALF, i2_ },
                                            fset.ex1(x_Ph_H0));
-              } else if constexpr (traits::fieldsetter::HasDx1<I, Dim::_2D> and
+              } else if constexpr (::traits::fieldsetter::HasDx1<I, Dim::_2D> and
                                    S == SimEngine::GRPIC) {
                 Fld(i1, i2, em::dx1) = s * Fld(i1, i2, em::dx1) +
                                        (ONE - s) * fset.dx1(x_Ph_H0);
@@ -245,7 +247,7 @@ namespace kernel::bc {
             }
           }
 
-          if constexpr (traits::fieldsetter::HasBx2<I, Dim::_2D>) {
+          if constexpr (::traits::fieldsetter::HasBx2<I, Dim::_2D>) {
             if (tags & BC::B) {
               if constexpr (S == SimEngine::SRPIC) {
                 Fld(i1, i2, em::bx2) = s * Fld(i1, i2, em::bx2) +
@@ -261,9 +263,9 @@ namespace kernel::bc {
           }
         }
 
-        if constexpr (traits::fieldsetter::HasEx2<I, Dim::_2D> or
-                      traits::fieldsetter::HasDx2<I, Dim::_2D> or
-                      traits::fieldsetter::HasBx1<I, Dim::_2D>) {
+        if constexpr (::traits::fieldsetter::HasEx2<I, Dim::_2D> or
+                      ::traits::fieldsetter::HasDx2<I, Dim::_2D> or
+                      ::traits::fieldsetter::HasBx1<I, Dim::_2D>) {
           // i1, i2 + 1/2
           real_t xi_Cd;
           if constexpr (o == in::x1) {
@@ -278,17 +280,17 @@ namespace kernel::bc {
           coord_t<Dim::_2D> x_Ph_0H { ZERO };
           metric.template convert<Crd::Cd, Crd::Ph>({ i1_, i2_ + HALF }, x_Ph_0H);
 
-          if constexpr (traits::fieldsetter::HasEx2<I, Dim::_2D> or
-                        traits::fieldsetter::HasDx2<I, Dim::_2D>) {
+          if constexpr (::traits::fieldsetter::HasEx2<I, Dim::_2D> or
+                        ::traits::fieldsetter::HasDx2<I, Dim::_2D>) {
             if ((tags & BC::E) or (tags & BC::D)) {
-              if constexpr (traits::fieldsetter::HasEx2<I, Dim::_2D> and
+              if constexpr (::traits::fieldsetter::HasEx2<I, Dim::_2D> and
                             S == SimEngine::SRPIC) {
                 Fld(i1, i2, em::ex2) = s * Fld(i1, i2, em::ex2) +
                                        (ONE - s) *
                                          metric.template transform<2, Idx::T, Idx::U>(
                                            { i1_, i2_ + HALF },
                                            fset.ex2(x_Ph_0H));
-              } else if constexpr (traits::fieldsetter::HasDx2<I, Dim::_2D> and
+              } else if constexpr (::traits::fieldsetter::HasDx2<I, Dim::_2D> and
                                    S == SimEngine::GRPIC) {
                 Fld(i1, i2, em::dx2) = s * Fld(i1, i2, em::dx2) +
                                        (ONE - s) * fset.dx2(x_Ph_0H);
@@ -296,7 +298,7 @@ namespace kernel::bc {
             }
           }
 
-          if constexpr (traits::fieldsetter::HasBx1<I, Dim::_2D>) {
+          if constexpr (::traits::fieldsetter::HasBx1<I, Dim::_2D>) {
             if (tags & BC::B) {
               if constexpr (S == SimEngine::SRPIC) {
                 Fld(i1, i2, em::bx1) = s * Fld(i1, i2, em::bx1) +
@@ -313,8 +315,8 @@ namespace kernel::bc {
         }
 
         // if constexpr (defines_ex3 or defines_dx3) {
-        if constexpr (traits::fieldsetter::HasEx3<I, Dim::_2D> or
-                      traits::fieldsetter::HasDx3<I, Dim::_2D>) {
+        if constexpr (::traits::fieldsetter::HasEx3<I, Dim::_2D> or
+                      ::traits::fieldsetter::HasDx3<I, Dim::_2D>) {
           if (tags & BC::E) {
             // i1, i2
             real_t xi_Cd;
@@ -330,7 +332,7 @@ namespace kernel::bc {
             coord_t<Dim::_2D> x_Ph_00 { ZERO };
             metric.template convert<Crd::Cd, Crd::Ph>({ i1_, i2_ }, x_Ph_00);
 
-            if constexpr (traits::fieldsetter::HasEx3<I, Dim::_2D> and
+            if constexpr (::traits::fieldsetter::HasEx3<I, Dim::_2D> and
                           S == SimEngine::SRPIC) {
               Fld(i1, i2, em::ex3) = s * Fld(i1, i2, em::ex3);
               if ((!is_axis_i2min or (i2 > N_GHOSTS)) and
@@ -340,7 +342,7 @@ namespace kernel::bc {
                                           { i1_, i2_ },
                                           fset.ex3(x_Ph_00));
               }
-            } else if constexpr (traits::fieldsetter::HasDx3<I, Dim::_2D> and
+            } else if constexpr (::traits::fieldsetter::HasDx3<I, Dim::_2D> and
                                  S == SimEngine::GRPIC) {
               Fld(i1, i2, em::dx3) = s * Fld(i1, i2, em::dx3);
               if ((!is_axis_i2min or (i2 > N_GHOSTS)) and
@@ -351,7 +353,7 @@ namespace kernel::bc {
           }
         }
 
-        if constexpr (traits::fieldsetter::HasBx3<I, Dim::_2D>) {
+        if constexpr (::traits::fieldsetter::HasBx3<I, Dim::_2D>) {
           if (tags & BC::B) {
             // i1 + 1/2, i2 + 1/2
             real_t xi_Cd;
@@ -395,11 +397,11 @@ namespace kernel::bc {
 
         if constexpr (S == SimEngine::SRPIC) {
           // SRPIC
-          if constexpr (traits::fieldsetter::HasEx1<I, Dim::_3D> or
-                        traits::fieldsetter::HasEx2<I, Dim::_3D> or
-                        traits::fieldsetter::HasEx3<I, Dim::_3D>) {
+          if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_3D> or
+                        ::traits::fieldsetter::HasEx2<I, Dim::_3D> or
+                        ::traits::fieldsetter::HasEx3<I, Dim::_3D>) {
             if (tags & BC::E) {
-              if constexpr (traits::fieldsetter::HasEx1<I, Dim::_3D>) {
+              if constexpr (::traits::fieldsetter::HasEx1<I, Dim::_3D>) {
                 // i1 + 1/2, i2, i3
                 real_t xi_Cd;
                 if constexpr (o == in::x1) {
@@ -423,7 +425,7 @@ namespace kernel::bc {
                                 fset.ex1(x_Ph_H00));
               }
 
-              if constexpr (traits::fieldsetter::HasEx2<I, Dim::_3D>) {
+              if constexpr (::traits::fieldsetter::HasEx2<I, Dim::_3D>) {
                 // i1, i2 + 1/2, i3
                 real_t xi_Cd;
                 if constexpr (o == in::x1) {
@@ -447,7 +449,7 @@ namespace kernel::bc {
                                 fset.ex2(x_Ph_0H0));
               }
 
-              if constexpr (traits::fieldsetter::HasEx3<I, Dim::_3D>) {
+              if constexpr (::traits::fieldsetter::HasEx3<I, Dim::_3D>) {
                 // i1, i2, i3 + 1/2
                 real_t xi_Cd;
                 if constexpr (o == in::x1) {
@@ -475,11 +477,11 @@ namespace kernel::bc {
             }
           }
 
-          if constexpr (traits::fieldsetter::HasBx1<I, Dim::_3D> or
-                        traits::fieldsetter::HasBx2<I, Dim::_3D> or
-                        traits::fieldsetter::HasBx3<I, Dim::_3D>) {
+          if constexpr (::traits::fieldsetter::HasBx1<I, Dim::_3D> or
+                        ::traits::fieldsetter::HasBx2<I, Dim::_3D> or
+                        ::traits::fieldsetter::HasBx3<I, Dim::_3D>) {
             if (tags & BC::B) {
-              if constexpr (traits::fieldsetter::HasBx1<I, Dim::_3D>) {
+              if constexpr (::traits::fieldsetter::HasBx1<I, Dim::_3D>) {
                 // i1, i2 + 1/2, i3 + 1/2
                 real_t xi_Cd;
                 if constexpr (o == in::x1) {
@@ -504,7 +506,7 @@ namespace kernel::bc {
                                 fset.bx1(x_Ph_0HH));
               }
 
-              if constexpr (traits::fieldsetter::HasBx2<I, Dim::_3D>) {
+              if constexpr (::traits::fieldsetter::HasBx2<I, Dim::_3D>) {
                 // i1 + 1/2, i2, i3 + 1/2
                 real_t xi_Cd;
                 if constexpr (o == in::x1) {
@@ -529,7 +531,7 @@ namespace kernel::bc {
                                 fset.bx2(x_Ph_H0H));
               }
 
-              if constexpr (traits::fieldsetter::HasBx3<I, Dim::_3D>) {
+              if constexpr (::traits::fieldsetter::HasBx3<I, Dim::_3D>) {
                 // i1 + 1/2, i2 + 1/2, i3
                 real_t xi_Cd;
                 if constexpr (o == in::x1) {
@@ -971,12 +973,12 @@ namespace kernel::bc {
     requires metric::traits::HasD<M> && metric::traits::HasTransform_i<M> &&
              metric::traits::HasConvert<M> &&
              (static_cast<dim_t>(O) < static_cast<dim_t>(M::Dim)) &&
-             (traits::fieldsetter::HasEx1<I, M::Dim> ||
-              traits::fieldsetter::HasEx2<I, M::Dim> ||
-              traits::fieldsetter::HasEx3<I, M::Dim> ||
-              traits::fieldsetter::HasBx1<I, M::Dim> ||
-              traits::fieldsetter::HasBx2<I, M::Dim> ||
-              traits::fieldsetter::HasBx3<I, M::Dim>)
+             (::traits::fieldsetter::HasEx1<I, M::Dim> ||
+              ::traits::fieldsetter::HasEx2<I, M::Dim> ||
+              ::traits::fieldsetter::HasEx3<I, M::Dim> ||
+              ::traits::fieldsetter::HasBx1<I, M::Dim> ||
+              ::traits::fieldsetter::HasBx2<I, M::Dim> ||
+              ::traits::fieldsetter::HasBx3<I, M::Dim>)
   struct EnforcedBoundaries_kernel {
     static constexpr Dimension D = M::Dim;
 
@@ -1004,12 +1006,12 @@ namespace kernel::bc {
         coord_t<Dim::_1D> x_Ph_H { ZERO };
         metric.template convert<Crd::Cd, Crd::Ph>({ i1_ }, x_Ph_0);
         metric.template convert<Crd::Cd, Crd::Ph>({ i1_ + HALF }, x_Ph_H);
-        bool setEx1 = traits::fieldsetter::HasEx1<I, M::Dim> and (tags & BC::E),
-             setEx2 = traits::fieldsetter::HasEx2<I, M::Dim> and (tags & BC::E),
-             setEx3 = traits::fieldsetter::HasEx3<I, M::Dim> and (tags & BC::E),
-             setBx1 = traits::fieldsetter::HasBx1<I, M::Dim> and (tags & BC::B),
-             setBx2 = traits::fieldsetter::HasBx2<I, M::Dim> and (tags & BC::B),
-             setBx3 = traits::fieldsetter::HasBx3<I, M::Dim> and (tags & BC::B);
+        bool setEx1 = ::traits::fieldsetter::HasEx1<I, M::Dim> and (tags & BC::E),
+             setEx2 = ::traits::fieldsetter::HasEx2<I, M::Dim> and (tags & BC::E),
+             setEx3 = ::traits::fieldsetter::HasEx3<I, M::Dim> and (tags & BC::E),
+             setBx1 = ::traits::fieldsetter::HasBx1<I, M::Dim> and (tags & BC::B),
+             setBx2 = ::traits::fieldsetter::HasBx2<I, M::Dim> and (tags & BC::B),
+             setBx3 = ::traits::fieldsetter::HasBx3<I, M::Dim> and (tags & BC::B);
         if constexpr (O == in::x1) {
           // x1 -- normal
           // x2,x3 -- tangential
@@ -1025,42 +1027,42 @@ namespace kernel::bc {
         } else {
           raise::KernelError(HERE, "Invalid Orientation");
         }
-        if constexpr (traits::fieldsetter::HasEx1<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx1<I, M::Dim>) {
           if (setEx1) {
             Fld(i1, em::ex1) = metric.template transform<1, Idx::T, Idx::U>(
               { i1_ + HALF },
               fset.ex1(x_Ph_H));
           }
         }
-        if constexpr (traits::fieldsetter::HasEx2<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx2<I, M::Dim>) {
           if (setEx2) {
             Fld(i1, em::ex2) = metric.template transform<2, Idx::T, Idx::U>(
               { i1_ },
               fset.ex2(x_Ph_0));
           }
         }
-        if constexpr (traits::fieldsetter::HasEx3<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx3<I, M::Dim>) {
           if (setEx3) {
             Fld(i1, em::ex3) = metric.template transform<3, Idx::T, Idx::U>(
               { i1_ },
               fset.ex3(x_Ph_0));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx1<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx1<I, M::Dim>) {
           if (setBx1) {
             Fld(i1, em::bx1) = metric.template transform<1, Idx::T, Idx::U>(
               { i1_ },
               fset.bx1(x_Ph_0));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx2<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx2<I, M::Dim>) {
           if (setBx2) {
             Fld(i1, em::bx2) = metric.template transform<2, Idx::T, Idx::U>(
               { i1_ + HALF },
               fset.bx2(x_Ph_H));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx3<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx3<I, M::Dim>) {
           if (setBx3) {
             Fld(i1, em::bx3) = metric.template transform<3, Idx::T, Idx::U>(
               { i1_ + HALF },
@@ -1085,12 +1087,12 @@ namespace kernel::bc {
         metric.template convert<Crd::Cd, Crd::Ph>({ i1_ + HALF, i2_ }, x_Ph_H0);
         metric.template convert<Crd::Cd, Crd::Ph>({ i1_ + HALF, i2_ + HALF },
                                                   x_Ph_HH);
-        bool setEx1 = traits::fieldsetter::HasEx1<I, M::Dim> and (tags & BC::E),
-             setEx2 = traits::fieldsetter::HasEx2<I, M::Dim> and (tags & BC::E),
-             setEx3 = traits::fieldsetter::HasEx3<I, M::Dim> and (tags & BC::E),
-             setBx1 = traits::fieldsetter::HasBx1<I, M::Dim> and (tags & BC::B),
-             setBx2 = traits::fieldsetter::HasBx2<I, M::Dim> and (tags & BC::B),
-             setBx3 = traits::fieldsetter::HasBx3<I, M::Dim> and (tags & BC::B);
+        bool setEx1 = ::traits::fieldsetter::HasEx1<I, M::Dim> and (tags & BC::E),
+             setEx2 = ::traits::fieldsetter::HasEx2<I, M::Dim> and (tags & BC::E),
+             setEx3 = ::traits::fieldsetter::HasEx3<I, M::Dim> and (tags & BC::E),
+             setBx1 = ::traits::fieldsetter::HasBx1<I, M::Dim> and (tags & BC::B),
+             setBx2 = ::traits::fieldsetter::HasBx2<I, M::Dim> and (tags & BC::B),
+             setBx3 = ::traits::fieldsetter::HasBx3<I, M::Dim> and (tags & BC::B);
 
         if constexpr (O == in::x1) {
           // x1 -- normal
@@ -1119,42 +1121,42 @@ namespace kernel::bc {
         } else {
           raise::KernelError(HERE, "Invalid Orientation");
         }
-        if constexpr (traits::fieldsetter::HasEx1<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx1<I, M::Dim>) {
           if (setEx1) {
             Fld(i1, i2, em::ex1) = metric.template transform<1, Idx::T, Idx::U>(
               { i1_ + HALF, i2_ },
               fset.ex1(x_Ph_H0));
           }
         }
-        if constexpr (traits::fieldsetter::HasEx2<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx2<I, M::Dim>) {
           if (setEx2) {
             Fld(i1, i2, em::ex2) = metric.template transform<2, Idx::T, Idx::U>(
               { i1_, i2_ + HALF },
               fset.ex2(x_Ph_0H));
           }
         }
-        if constexpr (traits::fieldsetter::HasEx3<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx3<I, M::Dim>) {
           if (setEx3) {
             Fld(i1, i2, em::ex3) = metric.template transform<3, Idx::T, Idx::U>(
               { i1_, i2_ },
               fset.ex3(x_Ph_00));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx1<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx1<I, M::Dim>) {
           if (setBx1) {
             Fld(i1, i2, em::bx1) = metric.template transform<1, Idx::T, Idx::U>(
               { i1_, i2_ + HALF },
               fset.bx1(x_Ph_0H));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx2<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx2<I, M::Dim>) {
           if (setBx2) {
             Fld(i1, i2, em::bx2) = metric.template transform<2, Idx::T, Idx::U>(
               { i1_ + HALF, i2_ },
               fset.bx2(x_Ph_H0));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx3<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx3<I, M::Dim>) {
           if (setBx3) {
             Fld(i1, i2, em::bx3) = metric.template transform<3, Idx::T, Idx::U>(
               { i1_ + HALF, i2_ + HALF },
@@ -1190,12 +1192,12 @@ namespace kernel::bc {
                                                   x_Ph_H0H);
         metric.template convert<Crd::Cd, Crd::Ph>({ i1_, i2_ + HALF, i3_ + HALF },
                                                   x_Ph_0HH);
-        bool setEx1 = traits::fieldsetter::HasEx1<I, M::Dim> and (tags & BC::E),
-             setEx2 = traits::fieldsetter::HasEx2<I, M::Dim> and (tags & BC::E),
-             setEx3 = traits::fieldsetter::HasEx3<I, M::Dim> and (tags & BC::E),
-             setBx1 = traits::fieldsetter::HasBx1<I, M::Dim> and (tags & BC::B),
-             setBx2 = traits::fieldsetter::HasBx2<I, M::Dim> and (tags & BC::B),
-             setBx3 = traits::fieldsetter::HasBx3<I, M::Dim> and (tags & BC::B);
+        bool setEx1 = ::traits::fieldsetter::HasEx1<I, M::Dim> and (tags & BC::E),
+             setEx2 = ::traits::fieldsetter::HasEx2<I, M::Dim> and (tags & BC::E),
+             setEx3 = ::traits::fieldsetter::HasEx3<I, M::Dim> and (tags & BC::E),
+             setBx1 = ::traits::fieldsetter::HasBx1<I, M::Dim> and (tags & BC::B),
+             setBx2 = ::traits::fieldsetter::HasBx2<I, M::Dim> and (tags & BC::B),
+             setBx3 = ::traits::fieldsetter::HasBx3<I, M::Dim> and (tags & BC::B);
         if constexpr (O == in::x1) {
           // x1 -- normal
           // x2,x3 -- tangential
@@ -1235,42 +1237,42 @@ namespace kernel::bc {
         } else {
           raise::KernelError(HERE, "Invalid Orientation");
         }
-        if constexpr (traits::fieldsetter::HasEx1<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx1<I, M::Dim>) {
           if (setEx1) {
             Fld(i1, i2, i3, em::ex1) = metric.template transform<1, Idx::T, Idx::U>(
               { i1_ + HALF, i2_, i3_ },
               fset.ex1(x_Ph_H00));
           }
         }
-        if constexpr (traits::fieldsetter::HasEx2<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx2<I, M::Dim>) {
           if (setEx2) {
             Fld(i1, i2, i3, em::ex2) = metric.template transform<2, Idx::T, Idx::U>(
               { i1_, i2_ + HALF, i3_ },
               fset.ex2(x_Ph_0H0));
           }
         }
-        if constexpr (traits::fieldsetter::HasEx3<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasEx3<I, M::Dim>) {
           if (setEx3) {
             Fld(i1, i2, i3, em::ex3) = metric.template transform<3, Idx::T, Idx::U>(
               { i1_, i2_, i3_ + HALF },
               fset.ex3(x_Ph_00H));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx1<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx1<I, M::Dim>) {
           if (setBx1) {
             Fld(i1, i2, i3, em::bx1) = metric.template transform<1, Idx::T, Idx::U>(
               { i1_, i2_ + HALF, i3_ + HALF },
               fset.bx1(x_Ph_0HH));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx2<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx2<I, M::Dim>) {
           if (setBx2) {
             Fld(i1, i2, i3, em::bx2) = metric.template transform<2, Idx::T, Idx::U>(
               { i1_ + HALF, i2_, i3_ + HALF },
               fset.bx2(x_Ph_H0H));
           }
         }
-        if constexpr (traits::fieldsetter::HasBx3<I, M::Dim>) {
+        if constexpr (::traits::fieldsetter::HasBx3<I, M::Dim>) {
           if (setBx3) {
             Fld(i1, i2, i3, em::bx3) = metric.template transform<3, Idx::T, Idx::U>(
               { i1_ + HALF, i2_ + HALF, i3_ },
